% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_bounds.r
\name{all_bounds}
\alias{all_bounds}
\title{Bounds for all coefficients}
\usage{
all_bounds(YL, YU, X.)
}
\arguments{
\item{YL}{response lower bound}

\item{YU}{response upper bound}

\item{X.}{design matrix}
}
\value{
matrix containing the estimated bounds
}
\description{
Bounds for all coefficients
}
\author{
Brenton Kenkel
}
\keyword{internal}

