% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_bounds.r
\name{boot_dist}
\alias{boot_dist}
\title{Calculate bootstrap distances}
\usage{
boot_dist(fit_main, fit_boot, n_obs)
}
\arguments{
\item{fit_main}{Matrix of sample bound estimates}

\item{fit_boot}{List of matrices of bootstrap bound estimates, as returned
by boot_coefs()}

\item{n_obs}{Number of observations}
}
\value{
List of boot x 2 matrices of bootstrap Hausdorff distances
    multiplied by root-N (one list element per term, containing directed and
    undirected distances)
}
\description{
Calculate bootstrap distances
}
\author{
Brenton Kenkel
}
\keyword{internal}

