% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_crit.r
\name{calculate_crit}
\alias{calculate_crit}
\title{Calculate critical values from bootstrap results}
\usage{
calculate_crit(fit, term, level = 0.95, directed = TRUE)
}
\arguments{
\item{fit}{Object of class \code{"coefbounds"}}

\item{term}{Index or name of term in \code{fit}}

\item{level}{Confidence level}

\item{directed}{Whether to calculate the critical value for the directed or
undirected Hausdorff distances}
}
\value{
Critical value for the given level
}
\description{
Calculate critical values from bootstrap results
}
\author{
Brenton Kenkel
}
\keyword{internal}

