% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_bounds.r
\name{coef_bounds}
\alias{coef_bounds}
\title{Bounds for a single coefficient}
\usage{
coef_bounds(d, YL, YU, X.)
}
\arguments{
\item{d}{column of design matrix to estimate bounds for}

\item{YL}{response lower bound}

\item{YU}{response upper bound}

\item{X.}{design matrix}
}
\value{
numeric vector containing the estimated bounds
}
\description{
Bounds for a single coefficient
}
\author{
Brenton Kenkel
}
\keyword{internal}

