% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefbounds.r
\name{coefbounds}
\alias{coefbounds}
\title{Coefficient bounds for linear models}
\usage{
coefbounds(formula, data, subset, na.action, model = c("linear", "logit"),
  boot = 100, maxit = 10, remove_collinear = TRUE,
  return_boot_est = FALSE)
}
\arguments{
\item{formula}{Model formula of the form \code{yl + yu ~ x1 + x2 + ...},
where \code{yl} is the lower bound on the response, \code{yu} is the
upper bound, and \code{x1 + x2 + ...} are the covariates.  For
instrumental variables estimation, use a formula like \code{yl + yu ~ x1
+ x2 + ... | z1 + z2 + ...}, as in \code{\link[AER]{ivreg}} in the
\pkg{AER} package.  IV estimates not available for logit models.}

\item{data, subset, na.action}{As in \code{\link{lm}}}

\item{model}{\code{"linear"} for linear regression (default), \code{"logit"}
for logistic regression.}

\item{boot}{Number of bootstrap iterations used to estimate the critical
values for inference.}

\item{maxit}{Maximum number of iterations for the approximation in logistic
regression models.  Ignored when \code{model = "linear"}.}

\item{remove_collinear}{How to treat boostrap iterations in which the design
matrix is rank-deficient.  If \code{TRUE} (the default), a warning is
issued and the bad iterations are removed.  If \code{FALSE}, the
function fails with an error when a rank-deficient design matrix is
encountered.}

\item{return_boot_est}{Whether to include the bootstrap estimates of the
coefficient bounds in the returned object.}
}
\value{
A list of class \code{"coefbounds"} containing: \describe{
\item{\code{coefficients}}{Matrix containing the sample estimates of the
    coefficient bounds.}
\item{\code{dist}}{List of matrices containing the
    bootstrap Hausdorff distances (undirected and directed) used for
    inference.}
\item{\code{boot_est}}{(if requested) List of matrices of
    bootstrap estimates of the coefficient bounds.}
\item{\code{nobs}}{Number of observations used in fitting.}
\item{\code{call}}{Original function call.}
\item{\code{model}}{Model used.}
}
}
\description{
Estimates the projections of the identification region along each
coefficient dimension for a linear or logistic regression model with
interval-censored outcomes (Beresteanu and Molinari 2008, Corollary 4.5).
If requested, uses a nonparametric bootstrap to estimate critical values for
hypothesis tests about these projections (Beresteanu and Molinari 2008,
Algorithm 4.2).
}
\details{
In the linear case, implements largely the same functionality as
    \code{oneDproj} and \code{CI1D} in Beresteanu et al.'s (2010) Stata
    program.
}
\examples{
## Simulate data
set.seed(18)
x1 <- rnorm(50)
x2 <- rnorm(50)
y <- 1 - x1 + x2 + rnorm(50)
yl <- floor(y)
yu <- ceiling(y)

## Fit model without covariates
fit_mean <- coefbounds(yl + yu ~ 1, boot = 0)
all.equal(coef(fit_mean)[1, "lower"], mean(yl))
all.equal(coef(fit_mean)[1, "upper"], mean(yu))

## Fit model with covariates
fit_full <- coefbounds(yl + yu ~ x1 + x2, boot = 10)
coef(fit_full)
}
\author{
Brenton Kenkel
}
\references{
Arie Beresteanu and Francesca Molinari.  2008.
    "Asymptotic Properties for a Class of Partially Identified Models."
    \emph{Econometrica} 76 (4): 763--814.

Arie Beresteanu, Francesca Molinari and Darcy Steeg Morris.  2010.
    "Asymptotics for Partially Identified Models in Stata."
    \url{https://molinari.economics.cornell.edu/programs.html}
}

