% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.coefbounds.r
\name{confint.coefbounds}
\alias{confint.coefbounds}
\title{Confidence regions for coefficient bounds}
\usage{
\method{confint}{coefbounds}(object, parm = NULL, level = 0.95,
  type = c("DU", "CC"), ...)
}
\arguments{
\item{object}{Fitted model of class \code{"coefbounds"}.  Must have been run
with \code{boot > 0} and contain a non-null \code{dist} element.}

\item{parm}{Names or indices of the coefficients for which to estimate
confidence sets.  If not specified, uses all coefficients.}

\item{level}{Confidence level.}

\item{type}{Whether to return the DU confidence set or the CC confidence
collections.  See Details for explanation.}

\item{...}{Not used.}
}
\value{
A matrix containing the range of the confidence region for each term
    requested.  When \code{type== "CC"}, there are two entries per term
    (range of the confidence region for the lower and upper bound).
}
\description{
Bootstrap estimates of confidence regions for the bounds on model
coefficients.
}
\details{
Two types of confidence region are available, both defined in Beresteanu and
Molinari (2008, Section 2).  The default is \code{"DU"}, the confidence
region based on the directed Hausdorff distance.  We would fail to reject
the null hypothesis that \eqn{B} is a subset of the population
identification region for the coefficient \eqn{\beta_j} if and only if
\eqn{B \subseteq DU(\beta_j)}.  The other option is \code{"CC"}, the
confidence collection based on the undirected Hausdorff distance.  We would
fail to reject the null hypothesis that \eqn{[b_L, b_U]} equals the
population identification region for \eqn{\beta_j} if and only if \eqn{b_L
\in CC_L(\beta_j)} and \eqn{b_U \in CC_U(\beta_j)}.

The \code{"DU"} confidence region corresponds to
\code{\link{interval_hypothesis}} with \code{type = "subset"}.  The
\code{"CC"} confidence region corresponds to
\code{\link{interval_hypothesis}} with \code{type = "equal"}.

Implements largely the same functionality as \code{CI1D} in Beresteanu et
    al.'s (2010) Stata program.
}
\examples{
## Simulate data
set.seed(18)
x1 <- rnorm(50)
x2 <- rnorm(50)
y <- 1 - x1 + x2 + rnorm(50)
yl <- floor(y)
yu <- ceiling(y)

## Fit model
fit <- lm_bounds(yl + yu ~ x1 + x2, boot = 100)

## Calculate DU confidence region for coefficient on x1
confint(fit, parm = "x1")

## Calculate CC confidence collections for all coefficients
confint(fit, type = "CC")
}
\author{
Brenton Kenkel
}
\references{
Arie Beresteanu and Francesca Molinari.  2008.
    "Asymptotic Properties for a Class of Partially Identified Models."
    \emph{Econometrica} 76 (4): 763--814.

Arie Beresteanu, Francesca Molinari and Darcy Steeg Morris.  2010.
    "Asymptotics for Partially Identified Models in Stata."
    \url{https://molinari.economics.cornell.edu/programs.html}
}

