% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.r
\name{draw_boot_ind}
\alias{draw_boot_ind}
\title{Draw indices for cluster bootstrap}
\usage{
draw_boot_ind(cluster_id)
}
\arguments{
\item{cluster_id}{Vector of cluster IDs for cluster bootstrap.  If
\code{NULL} (the default), an ordinary bootstrap is used.}
}
\value{
vector of indices
}
\description{
Draw indices for cluster bootstrap
}
\author{
Brenton Kenkel
}
\keyword{internal}

