% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.coefbounds.r
\name{format.perc}
\alias{format.perc}
\title{stats:::format.perc}
\usage{
\method{format}{perc}(probs, digits)
}
\arguments{
\item{probs}{probabilities}

\item{digits}{number of digits}
}
\value{
(kinda) pretty names, consistent with those used in the default
    confint() method
}
\description{
stats:::format.perc
}
\keyword{internal}

