% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hausdorff_distance.r
\name{hausdorff_distance}
\alias{hausdorff_distance}
\title{Hausdorff distance between intervals}
\usage{
hausdorff_distance(x, y, directed = TRUE)
}
\arguments{
\item{x, y}{2-column matrices, where the first column is pointwise less than
or equal to the second}

\item{directed}{logical: calculate the directed distance?}
}
\value{
vector of distances
}
\description{
Calculates the directed or undirected Hausdorff distance between pairs of
intervals.
}
\author{
Brenton Kenkel
}
\keyword{internal}

