% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_bounds.r
\name{lm_bounds}
\alias{lm_bounds}
\title{Bounds for all coefficients in a linear model}
\usage{
lm_bounds(YL, YU, X, Z, ...)
}
\arguments{
\item{YL}{response lower bound}

\item{YU}{response upper bound}

\item{X}{design matrix}

\item{Z}{instrument matrix}

\item{...}{for safety}
}
\value{
matrix containing the estimated bounds
}
\description{
Bounds for all coefficients in a linear model
}
\author{
Brenton Kenkel
}
\keyword{internal}

