% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_bounds.r
\name{logit_bounds}
\alias{logit_bounds}
\title{Bounds for all coefficients in a logit model}
\usage{
logit_bounds(YL, YU, X, maxit, ...)
}
\arguments{
\item{YL}{response lower bound}

\item{YU}{response upper bound}

\item{X}{design matrix}

\item{maxit}{max iterations for logit approximation}

\item{...}{for safety}
}
\value{
matrix containing the estimated bounds
}
\description{
Bounds for all coefficients in a logit model
}
\author{
Brenton Kenkel
}
\keyword{internal}

