% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_bounds.r
\name{logit_coefs}
\alias{logit_coefs}
\title{Logit coefficients and variance matrix}
\usage{
logit_coefs(Y, X)
}
\arguments{
\item{Y}{response}

\item{X}{design matrix}
}
\value{
list containing the coefficients and variance matrix
}
\description{
Logit coefficients and variance matrix
}
\author{
Brenton Kenkel
}
\keyword{internal}

