% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hausdorff_distance.r
\name{point_to_interval_distance}
\alias{point_to_interval_distance}
\title{Distance from a point to an interval}
\usage{
point_to_interval_distance(point, interval)
}
\arguments{
\item{point}{vector of points}

\item{interval}{matrix of intervals, with 2 columns and number of rows equal
to the length of \code{point}}
}
\value{
vector of distances
}
\description{
Calculates the distance from each point to the closest element of the
corresponding interval.  Doesn't check validity of arguments; should only be
called within hausdorff_distance().
}
\author{
Brenton Kenkel
}
\keyword{internal}

