% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.r
\name{boot_coefs}
\alias{boot_coefs}
\title{Bootstrap coefficient bounds}
\usage{
boot_coefs(boot, remove_collinear, YL, YU, X, model, maxit)
}
\arguments{
\item{boot}{number of bootstrap iterations}

\item{remove_collinear}{if \code{TRUE}, removes bootstrap iterations with a
collinear design matrix (with warning); otherwise, any collinearity
triggers a failure}

\item{YL}{response lower bound}

\item{YU}{response upper bound}

\item{X}{design matrix}

\item{model}{"linear" or "logit"}

\item{maxit}{max iterations for logit approximation}
}
\value{
List of boot x 2 matrices of bootstrap bound estimates (one list
    element per term)
}
\description{
Bootstrap coefficient bounds
}
\author{
Brenton Kenkel
}
\keyword{internal}

